versionInfoCard component
